Block parameter on read is renamed to wait
Due to the presence of mutexes, it is possible for the "nonblocking"
version of reads to block. Therefore, the semantics of the method have
less to do with blocking or nonblocking, and more to do with whether or
not the user wishes to wait until there is a message to read.